chore: Delete `Readers.kt`
- `TrackGroupArrayException` will never be thrown, so we can remove it.
