--- android/src/main/java/com/cyanchill/missingcore/metadataretriever/Exceptions.kt
@@ -1,5 +1,0 @@
-package com.cyanchill.missingcore.metadataretriever
-
-
-class TrackGroupArrayException:
-  Exception("No `TrackGroupArray` Found Error")
